created separate file for SearchBar component
