+++ my-app/src/components/Recipe.js
@@ -3,54 +3,12 @@ import axiosWithAuth  from "./utils/AxiosWithAuth";
 import { useParams } from 'react-router-dom';
 import RecipesList from "./RecipesList";
 import AddRecipe from './AddRecipe';
+import SearchBar from './SearchBar';
 
 const Recipe = () => {
 
   const [recipeList, setRecipeList] = useState([]);
   const { id } = useParams(); 
-  // const [title, setTitle] = useState('');
-  const [searchTerm, setSearchTerm] = useState('');
-  const [unfilteredRecipes, setUnfitleredRecipes] = useState(recipeList);
-  
- 
-
-  const filterPosts =() =>{
-    return unfilteredRecipes.filter(recipe=>{
-      if(searchTerm=== '' ){
-        return axiosWithAuth()
-        .get(`/${id}/recipe/title?title=${searchTerm}`)
-      }
-    if(recipe.title.includes(searchTerm) || recipe.category.includes(searchTerm))
-    {
-      return recipe
-    }
-  })
-}
-
-const changeHandler = e =>{
-  const searchTerm = e.target.value
-  setSearchTerm(searchTerm);
-}
-
-const SearchBar = ()=>{
-
-
-  return(
-    <div className="searchBar">
-      <form>
-      <input
-       onChange ={changeHandler}
-       type="text"
-       placeholder="search"
-       value ={searchTerm}
-       />
-   
-      </form>
-    </div>
-  )
-}
-  
-
 
 
   useEffect(() => {
@@ -66,7 +24,7 @@ const SearchBar = ()=>{
   return (
     <div>
       <h1>Welcome to Secret Family Recipes!</h1>
-      <SearchBar/>
+      <SearchBar recipes={recipeList}/>
       <AddRecipe/>
       <RecipesList recipes={recipeList} />
     </div>
